requirements.txt
# requirements.txt
mcp>=0.1.0  # or fastmcp>=2.0.0 for easier development
pyiceberg==0.7.1  # Latest PyIceberg that supports Iceberg 1.7.0
pydantic>=2.0
python-dotenv  # for configuration
typing-extensions>=4.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.18.0  # For testing async functions
pytest-cov>=4.0.0  # For test coverage
pytest-mock>=3.10.0  # For mocking
